+++ Client/src/main/java/de/baleipzig/pdfextraction/client/connector/api/TemplateConnector.java
@@ -1,7 +1,6 @@
 package de.baleipzig.pdfextraction.client.connector.api;
 
 import de.baleipzig.pdfextraction.api.dto.TemplateDTO;
-import javafx.scene.image.Image;
 import reactor.core.publisher.Flux;
 import reactor.core.publisher.Mono;
 
@@ -29,22 +28,4 @@ public interface TemplateConnector {
      * @return empty Mono or error Mono
      */
     Mono<Void> save(TemplateDTO dto);
-
-    /**
-     * Actually executes an extraction of the PDF with the given template
-     *
-     * @param templateName Name of the Template to use
-     * @param pathToFile   Path to the PDF to use
-     * @return Map of fieldtype-names to their extracted Result
-     */
-    Mono<Map<String, String>> runJob(final String templateName, final Path pathToFile);
-
-    /**
-     * Creates an Image of the first Page of the given PDF with the Boxes of the given Template drawn to check if the Template matches the PDF
-     *
-     * @param templateName Name of the template to draw
-     * @param pathToFile   Path to the PDF to use
-     * @return Image
-     */
-    Mono<Image> createTestImage(final String templateName, final Path pathToFile);
 }
