+++ Client/src/main/java/de/baleipzig/pdfextraction/client/connector/impl/TemplateConnectorImpl.java
@@ -1,7 +1,7 @@
 package de.baleipzig.pdfextraction.client.connector.impl;
 
-import de.baleipzig.pdfextraction.api.config.Config;
 import de.baleipzig.pdfextraction.api.dto.TemplateDTO;
+import de.baleipzig.pdfextraction.client.config.Config;
 import de.baleipzig.pdfextraction.client.connector.api.TemplateConnector;
 import de.baleipzig.pdfextraction.client.utils.injector.ImplementationOrder;
 import org.springframework.core.ParameterizedTypeReference;
