+++ Client/src/main/java/de/baleipzig/pdfextraction/client/controller/PdfPreviewController.java
@@ -113,7 +113,7 @@ public class PdfPreviewController implements Initializable {
 
     public void updatePdfPreview() {
         // der restliche code wird in MenuBarController ausgeführt
-        if (this.renderer.hasPreview()){
+        if (this.renderer.hasPreview()) {
             loadPdfPreview(this.renderer::getCurrentPreview);
         }
     }
