+++ Client/src/main/resources/de/baleipzig/pdfextraction/client/view/Actions.fxml
@@ -6,7 +6,7 @@
 <?import javafx.geometry.Insets?>
 <?import javafx.scene.control.Label?>
 <?import javafx.scene.layout.*?>
-<BorderPane xmlns="http://javafx.com/javafx/17" xmlns:fx="http://javafx.com/fxml/1"
+<BorderPane xmlns:fx="http://javafx.com/fxml/1" xmlns="http://javafx.com/javafx/17"
             fx:controller="de.baleipzig.pdfextraction.client.controller.ActionController"
             stylesheets="@css/ViewStylesheet.css">
     <top>
