+++ Client/src/main/resources/de/baleipzig/pdfextraction/client/view/MenuBar.fxml
@@ -1,18 +1,16 @@
 <?xml version="1.0" encoding="UTF-8"?>
 
-<?import javafx.scene.control.Menu?>
-<?import javafx.scene.control.MenuBar?>
-<?import javafx.scene.control.MenuItem?>
-
-<MenuBar fx:id="menuBar" xmlns="http://javafx.com/javafx/17" xmlns:fx="http://javafx.com/fxml/1"
+<?import javafx.scene.control.*?>
+<MenuBar xmlns:fx="http://javafx.com/fxml/1" fx:id="menuBar" xmlns="http://javafx.com/javafx/17"
          fx:controller="de.baleipzig.pdfextraction.client.controller.MenuBarController">
-  <Menu mnemonicParsing="false" text="File">
-    <MenuItem fx:id="chooseFile" mnemonicParsing="false" onAction="#onChooseFile" text="Datei Auswählen"/>
-  </Menu>
-  <Menu mnemonicParsing="false" text="Template">
-    <MenuItem fx:id="createTemplate" mnemonicParsing="false" onAction="#onCreateTemplate" text="Template erstellen"/>
-  </Menu>
-  <Menu mnemonicParsing="false" text="Help">
-    <MenuItem mnemonicParsing="false" text="About"/>
-  </Menu>
+    <Menu mnemonicParsing="false" text="File">
+        <MenuItem fx:id="chooseFile" mnemonicParsing="false" onAction="#onChooseFile" text="Datei Auswählen"/>
+    </Menu>
+    <Menu mnemonicParsing="false" text="Template">
+        <MenuItem fx:id="createTemplate" mnemonicParsing="false" onAction="#onCreateTemplate"
+                  text="Template erstellen"/>
+    </Menu>
+    <Menu mnemonicParsing="false" text="Help">
+        <MenuItem mnemonicParsing="false" text="About"/>
+    </Menu>
 </MenuBar>
