+++ Client/src/test/java/de/baleipzig/pdfextraction/client/utils/TestExtractionConnector.java
@@ -1,7 +1,8 @@
-package de.baleipzig.pdfextraction.client.util;
+package de.baleipzig.pdfextraction.client.utils;
 
 import de.baleipzig.pdfextraction.client.connector.api.ExtractionConnector;
 import de.baleipzig.pdfextraction.client.utils.injector.ImplementationOrder;
+import javafx.scene.image.Image;
 import reactor.core.publisher.Mono;
 
 import java.nio.file.Path;
@@ -15,7 +16,7 @@ public class TestExtractionConnector implements ExtractionConnector {
     }
 
     @Override
-    public Mono<byte[]> createTestImage(String templateName, Path pathToFile) {
+    public Mono<Image> createTestImage(String templateName, Path pathToFile) {
         return Mono.error(new UnsupportedOperationException("Not implemented."));
     }
 }
